backend/requirements.txt
asgiref==3.8.1
Django==3.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
pytz==2024.1
sqlparse==0.5.0
psycopg2-binary==2.9.3
Pillow==9.3.0
requests==2.26.0
gunicorn==20.1.0
python-dotenv==1.0.1
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0